backend/requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==25.1
djangorestframework==3.14.0
Pillow==10.4.0
python-dotenv==1.0.0
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'
whitenoise==6.6.0

# Security packages
bleach==6.1.0
django-csp==3.8
django-permissions-policy==4.18.0


# Deployment dependencies
gunicorn==21.2.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# API and external services
google-generativeai==0.3.2
requests==2.31.0